Fix cut off notification sounds.
When posting a notification without sound, NotificationManagerService
now stops playing the sound of the previous notification. As we post an
update to the notification ~2s after we originally posted it, this leads
to cut off notification sounds.

Solution: If we only posted an update and didn't notify of new messages,
keep the sound URI set and instruct the notification manager to play the
sound only once.

JIRA: BUGBASH-199
Change-Id: I3aefa62ce1583839ef991380b2be6c39864509a1
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>